Optimize input code
