+++ src/components/index.ts
@@ -9,7 +9,8 @@ export { Card } from "./Card/Card";
 export { ContextMenu } from "./ContextMenu/ContextMenu";
 export { HoverCard } from "./HoverCard/HoverCard";
 export { Icon } from "./Icon/Icon";
-export { Input } from "./FormField/Input";
+export { Input } from "./FormField/Input/Input";
+export { PasswordInput } from "./FormField/Input/PasswordInput";
 export { Alert, DangerAlert, InfoAlert, WarningAlert, SuccessAlert } from "./Alert/Alert";
 export { Checkbox } from "./Checkbox/Checkbox";
 export { IconButton } from "./IconButton/IconButton";
